src/bot/handlers/portfolio.py
"""
Portfolio Management Handlers

This module handles all portfolio-related commands:
- /portfolio - View portfolio with current P&L
- /portfolio add SYMBOL SHARES PRICE - Add position
- /portfolio remove SYMBOL - Remove position
- /portfolio update SYMBOL SHARES PRICE - Update position
"""

import logging
from typing import List
from telegram import Update
from telegram.ext import ContextTypes

from ..database.db import (
    get_db_context,
    get_or_create_user
)
from ..services.analysis_service import get_current_price
from ..utils.formatters import format_success, format_error, format_warning
from ..utils.validators import validate_stock_symbol, validate_shares, validate_price, parse_command_args

logger = logging.getLogger(__name__)


async def portfolio_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
    """
    Handle /portfolio command.
    
    Shows user's portfolio with current P&L.
    
    Usage:
        /portfolio - View portfolio
        /portfolio add SYMBOL SHARES PRICE - Add position
        /portfolio remove SYMBOL - Remove position
    """
    user_id = update.effective_user.id
    username = update.effective_user.username
    
    # Parse subcommand
    args = parse_command_args(update.message.text)
    
    if not args:
        # Show portfolio
        await show_portfolio(update, context)
        return
    
    subcommand = args[0].lower()
    
    if subcommand == "add":
        await portfolio_add(update, context, args[1:])
    elif subcommand == "remove":
        await portfolio_remove(update, context, args[1:])
    elif subcommand == "update":
        await portfolio_update(update, context, args[1:])
    else:
        await update.message.reply_text(
            format_error(
                f"Unknown subcommand: {subcommand}\n\n"
                f"Available commands:\n"
                f"• /portfolio - View portfolio\n"
                f"• /portfolio add SYMBOL SHARES PRICE\n"
                f"• /portfolio remove SYMBOL"
            )
        )


async def show_portfolio(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
    """Show user's portfolio."""
    user_id = update.effective_user.id
    
    await update.message.reply_text(
        "💼 *Portfolio Feature Coming Soon!*\n\n"
        "The portfolio tracking feature is currently under development.\n\n"
        "*Planned features:*\n"
        "• Track your stock positions\n"
        "• Real-time P&L calculation\n"
        "• Performance metrics\n"
        "• Transaction history\n"
        "• Portfolio diversification analysis\n\n"
        "*Available now:*\n"
        "• Watchlist - Track stocks you're interested in\n"
        "• Alerts - Get notified of price/indicator changes\n"
        "• Analysis - Detailed technical analysis\n\n"
        "Stay tuned for updates!",
        parse_mode='Markdown'
    )


async def portfolio_add(update: Update, context: ContextTypes.DEFAULT_TYPE, args: List[str]) -> None:
    """Add position to portfolio."""
    await update.message.reply_text(
        "💼 *Add Position*\n\n"
        "This feature is coming soon!\n\n"
        "In the meantime, use the watchlist feature:\n"
        "`/watchlist add SYMBOL`",
        parse_mode='Markdown'
    )


async def portfolio_remove(update: Update, context: ContextTypes.DEFAULT_TYPE, args: List[str]) -> None:
    """Remove position from portfolio."""
    await update.message.reply_text(
        "💼 *Remove Position*\n\n"
        "This feature is coming soon!",
        parse_mode='Markdown'
    )


async def portfolio_update(update: Update, context: ContextTypes.DEFAULT_TYPE, args: List[str]) -> None:
    """Update position in portfolio."""
    await update.message.reply_text(
        "💼 *Update Position*\n\n"
        "This feature is coming soon!",
        parse_mode='Markdown'
    )
