requirements_bot.txt
# Stock Market Analyzer Pro - Telegram Bot Dependencies
# Python 3.9+ required

# Core dependencies (from main requirements.txt)
yahooquery>=2.3.0
pandas>=1.5.0
numpy>=1.23.0
ta>=0.10.0

# Telegram Bot Framework
python-telegram-bot[job-queue]==20.7

# Database ORM
SQLAlchemy>=2.0.0
alembic>=1.12.0

# Environment Variables
python-dotenv>=1.0.0

# Task Scheduling
APScheduler>=3.10.0

# Chart Generation
matplotlib>=3.8.0
mplfinance>=0.12.10b0
Pillow>=10.0.0

# Async HTTP requests
aiohttp>=3.9.0

# Caching
cachetools>=5.3.0

# Date/Time handling
python-dateutil>=2.8.0
pytz>=2023.3

# Rich console output (for bot logs)
rich>=13.0.0

# Development dependencies (optional)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
